requirements.txt
certifi==2019.6.16
Django==2.2.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.2
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
